isPolygon
