+++ packages/client/verifiable-credentials/src/types/nfts.ts
@@ -19,17 +19,3 @@ export interface CredentialsCollection {
         };
     };
 }
-
-export function parseLocator(locator: string): EVMNFT {
-    const items = locator.split(":");
-    const itemsLength = items.length;
-    if (itemsLength < 2) {
-        throw new Error(`Invalid locator format, expected <chain>:<contractAddress>:<tokenId>`);
-    }
-
-    return {
-        chain: items[0] as EVMBlockchain,
-        contractAddress: items[1],
-        tokenId: items[2],
-    };
-}
